pyproject.toml
[tool.poetry]
name = "pythonprofessional.2022-05.project"
version = "0.1.0"
description = ""
authors = ["Inna Sorina <sorina-inna@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
python-dotenv = "^0.21.0"
django-debug-toolbar = "^3.7.0"
psycopg2-binary = "^2.9.5"
requests = "^2.28.1"
celery = "^5.2.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
